feat: nav updated
